src/picture_diary/compose/export.py
from __future__ import annotations

from io import BytesIO

from PIL import Image


def image_to_png_bytes(image: Image.Image) -> bytes:
    buffer = BytesIO()
    image.save(buffer, format="PNG")
    return buffer.getvalue()
