requirements.txt
streamlit>=1.32
pillow>=10.2
python-dotenv>=1.0
huggingface_hub>=0.23
safetensors>=0.4
accelerate>=0.28
diffusers>=0.27
transformers>=4.39
peft>=0.11
torch>=2.1
torchvision>=0.18
openai>=1.14
pytest>=8.0
